Added in redux connect so data can be accessed through out the webapp
The reducers are the data that will be contained in the firestore datbase, for now it just consists of dummy data, i reckon the next step is to hook into the firestore and add sign in and sign out features.
Also addded functionality for scrollable lists on the todo page.
Its lookikng well