Implemented basic firebase functionality, adding to the database and also displaying the documents in the db
Also added buttons on each to do item. These buttons will be used to drigger the delete for completion of a task (if i delete them from the DB pr not is a different story havent decided yet)
Next step is to implement sign up and sign in ability  i reckon!